Broke up the Portfolio details into smaller components Changed the props of the shadow box for more background colours
