+++ src/pages/QuestradeOauthPage.js
@@ -110,7 +110,7 @@ const QuestradeOauthPage = ({ router, reloadAllState, push }) => {
   }
 
   return (
-    <ShadowBox dark>
+    <ShadowBox background="#2a2d34">
       <H1 color="white">SETUP</H1>
       {loading ? (
         <React.Fragment>
