+++ src/styled/ShadowBox.js
@@ -1,9 +1,13 @@
 import styled from '@emotion/styled';
+import { css } from '@emotion/core';
+
+const dynamicColor = props =>
+  css`
+    background: ${props.background};
+  `;
 
 export default styled.div`
   background: var(--white);
-  background-color: ${props =>
-    props.dark ? 'var(--brand-grey)' : 'var(--white)'};
   border-radius: 4px;
   box-shadow: var(--box-shadow);
   padding: 20px 20px 20px;
@@ -11,4 +15,5 @@ export default styled.div`
   @media (max-width: 900px) {
     padding: 15px;
   }
+  ${dynamicColor};
 `;
